requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0

sqlalchemy==2.0.29
psycopg2-binary==2.9.9

pydantic==2.6.4
pydantic-settings==2.2.1

alembic==1.13.1
python-dotenv>=1.0.1
requests>=2.31.0

# Testing
pytest==7.4.0
pytest-mock==3.14.0
httpx==0.27.0
anyio==4.2.0
coverage<7.5

# Logging enhancements (optional)
loguru==0.7.2
